RavenDB-24069 Move CanCreateRootAndManyBranchEnvironments to StressTests. Fixing warning message.
